+++ angular-workspace/projects/ni/nimble-angular/src/directives/anchor-tabs/nimble-anchor-tabs.module.ts
@@ -1,0 +1,12 @@
+import { NgModule } from '@angular/core';
+import { CommonModule } from '@angular/common';
+import { NimbleAnchorTabsDirective } from './nimble-anchor-tabs.directive';
+
+import '@ni/nimble-components/dist/esm/anchor-tabs';
+
+@NgModule({
+    declarations: [NimbleAnchorTabsDirective],
+    imports: [CommonModule],
+    exports: [NimbleAnchorTabsDirective]
+})
+export class NimbleAnchorTabsModule { }