+++ packages/nimble-components/src/table-column/text/tests/table-column-text.spec.ts
@@ -6,6 +6,8 @@ import { waitForUpdatesAsync } from '../../../testing/async-helpers';
 import { type Fixture, fixture } from '../../../utilities/tests/fixture';
 import type { TableRecord } from '../../../table/types';
 import { TablePageObject } from '../../../table/tests/table.pageobject';
+import { wackyStrings } from '../../../utilities/tests/wacky-strings';
+import { getSpecTypeByNamedList } from '../../../utilities/tests/parameterized';
 
 interface SimpleTableRecord extends TableRecord {
     field?: string | null;
@@ -52,7 +54,7 @@ describe('TableColumnText', () => {
     for (const testData of noValueData) {
         // eslint-disable-next-line @typescript-eslint/no-loop-func
         it(`displays placeholder string when ${testData.description}`, async () => {
-            element.data = testData.data;
+            element.setData(testData.data);
             await connect();
             await waitForUpdatesAsync();
 
@@ -61,7 +63,7 @@ describe('TableColumnText', () => {
     }
 
     it('changing fieldName updates display', async () => {
-        element.data = [{ field: 'foo', anotherField: 'bar' }];
+        element.setData([{ field: 'foo', anotherField: 'bar' }]);
         await connect();
         await waitForUpdatesAsync();
 
@@ -73,7 +75,7 @@ describe('TableColumnText', () => {
     });
 
     it('changing placeholder updates display', async () => {
-        element.data = [{ field: null }];
+        element.setData([{ field: null }]);
         await connect();
         await waitForUpdatesAsync();
 
@@ -85,26 +87,26 @@ describe('TableColumnText', () => {
     });
 
     it('changing data from value to null displays placeholder', async () => {
-        element.data = [{ field: 'foo' }];
+        element.setData([{ field: 'foo' }]);
         await connect();
         await waitForUpdatesAsync();
         expect(pageObject.getRenderedCellContent(0, 0)).toBe('foo');
 
         const updatedValue = { field: null };
         const updatedData = [updatedValue];
-        element.data = updatedData;
+        element.setData(updatedData);
         await waitForUpdatesAsync();
 
         expect(pageObject.getRenderedCellContent(0, 0)).toBe('no value');
     });
 
     it('changing data from null to value displays value', async () => {
-        element.data = [{ field: null }];
+        element.setData([{ field: null }]);
         await connect();
         await waitForUpdatesAsync();
         expect(pageObject.getRenderedCellContent(0, 0)).toBe('no value');
 
-        element.data = [{ field: 'foo' }];
+        element.setData([{ field: 'foo' }]);
         await waitForUpdatesAsync();
 
         expect(pageObject.getRenderedCellContent(0, 0)).toBe('foo');
@@ -116,29 +118,58 @@ describe('TableColumnText', () => {
 
         const firstColumn = element.columns[0] as TableColumnText;
         firstColumn.fieldName = undefined;
-        element.data = [{ field: 'foo' }];
+        element.setData([{ field: 'foo' }]);
         await waitForUpdatesAsync();
 
         expect(pageObject.getRenderedCellContent(0, 0)).toBe('');
     });
 
-    const fieldValues = [
-        { dataValue: 'foo', renderedValue: 'foo' },
-        { dataValue: '<button></button>', renderedValue: '<button></button>' },
-        { dataValue: 'null', renderedValue: 'null' },
-        { dataValue: 'undefined', renderedValue: 'undefined' }
-    ];
-    for (const fieldValue of fieldValues) {
-        // eslint-disable-next-line @typescript-eslint/no-loop-func
-        it(`data ${fieldValue.dataValue} renders as ${fieldValue.renderedValue}`, async () => {
-            await connect();
+    describe('various string values render as expected', () => {
+        const focused: string[] = [];
+        const disabled: string[] = [];
+        for (const value of wackyStrings) {
+            const specType = getSpecTypeByNamedList(value, focused, disabled);
+            // eslint-disable-next-line @typescript-eslint/no-loop-func
+            specType(
+                `data "${value.name}" renders as "${value.name}"`,
+                // eslint-disable-next-line @typescript-eslint/no-loop-func
+                async () => {
+                    await connect();
 
-            element.data = [{ field: fieldValue.dataValue }];
-            await waitForUpdatesAsync();
+                    element.setData([{ field: value.name }]);
+                    await waitForUpdatesAsync();
 
-            expect(pageObject.getRenderedCellContent(0, 0)).toBe(
-                fieldValue.renderedValue
+                    expect(pageObject.getRenderedCellContent(0, 0)).toBe(
+                        value.name
+                    );
+                }
             );
-        });
-    }
+        }
+    });
+
+    describe('placeholder assigned various strings render as expected', () => {
+        const focused: string[] = [];
+        const disabled: string[] = [];
+        for (const value of wackyStrings) {
+            const specType = getSpecTypeByNamedList(value, focused, disabled);
+            // eslint-disable-next-line @typescript-eslint/no-loop-func
+            specType(
+                `placeholder "${value.name}" renders as "${value.name}"`,
+                // eslint-disable-next-line @typescript-eslint/no-loop-func
+                async () => {
+                    await connect();
+                    element.setData([{ field: null }]);
+                    await waitForUpdatesAsync();
+
+                    const firstColumn = element.columns[0] as TableColumnText;
+                    firstColumn.placeholder = value.name;
+                    await waitForUpdatesAsync();
+
+                    expect(pageObject.getRenderedCellContent(0, 0)).toBe(
+                        value.name
+                    );
+                }
+            );
+        }
+    });
 });